Add MIT License on PressUP
This commit adds an MIT License to the PressUP repository to clearly define the permissions and limitations associated with the use of this project.

The MIT License allows for reuse, modification, and distribution, while limiting liability and warranty. Adding this license ensures the project is open-source and legally reusable by others.